src/api-gateway/requirements.txt
fastapi==0.115.4
httpx==0.27.2
PyYAML==6.0.2
uvicorn==0.32.0
